フォームの仕様の変更

+++ lib/component/uivalue.dart
@@ -168,7 +168,7 @@ class UIValue {
       {double defaultValue = 0, double filter(double value)}) {
     dynamic tmp = this.watch<dynamic>(path);
     if (tmp is double) return filter != null ? filter(tmp) : tmp;
-    if (tmp is int) return filter != null ? filter(tmp as double) : tmp as int;
+    if (tmp is int) return filter != null ? filter(tmp as double) : tmp;
     if (tmp is bool)
       return filter != null ? filter(tmp ? 1 : 0) : (tmp ? 1 : 0);
     if (tmp is String) {
